requirements.txt
git+https://github.com/farm-ng/farm-ng-core.git
grpcio-tools
kornia
kornia-rs
opencv-python
numpy